TagsPlugin: Prevent `AttributeError` on environment startup for Trac 0.13dev, refs #9521.
Additionally this broke all environment upgrades for plugins and for
Trac core itself, at least as long as the plugin was installed and enabled.

While this might be just a temporary Trac behavior (see Remy Blank's
[T:comment:ticket:10451:4 comment in ticket #10451]), it prevails in `trunk`
for more than a month now - a major headache for early adopters.
So I feel like acting now, facing the risk of not fixing it properly.

The logical next step is switching to the recommended procedure of storing
explicite db schema versions inside the Trac db like other plugins already do.

git-svn-id: http://trac-hacks.org/svn/tagsplugin/trunk@11041 7322e99d-02ea-0310-aa39-e9a107903beb